created external namespace for utility
